Move types to wrappers folder.
