+++ .gitignore
@@ -28,6 +28,7 @@ qmeq/approach/elph/c_lindblad.c
 qmeq/approach/elph/c_redfield.c
 qmeq/approach/elph/c_neumann1.c
 qmeq/wrappers/c_lapack.c
+qmeq/wrappers/c_mytypes.c
 
 ### Directories
