+++ qmeq/approach/base/lindblad.py
@@ -6,8 +6,8 @@ from __future__ import print_function
 import numpy as np
 import itertools
 
-from ...mytypes import complexnp
-from ...mytypes import doublenp
+from ...wrappers.mytypes import complexnp
+from ...wrappers.mytypes import doublenp
 
 from ...specfunc.specfunc import func_pauli
 from ..aprclass import Approach
