+++ qmeq/approach/c_aprclass.pxd
@@ -3,11 +3,11 @@ cimport numpy as np
 
 from ..wrappers.c_lapack cimport LapackSolver
 
-ctypedef np.uint8_t bool_t
-ctypedef np.int_t int_t
-ctypedef np.int64_t long_t
-ctypedef np.float64_t double_t
-ctypedef np.complex128_t complex_t
+from ..wrappers.c_mytypes cimport bool_t
+from ..wrappers.c_mytypes cimport int_t
+from ..wrappers.c_mytypes cimport long_t
+from ..wrappers.c_mytypes cimport double_t
+from ..wrappers.c_mytypes cimport complex_t
 
 
 cdef class Approach:
