+++ qmeq/approach/elph/c_pauli.pyx
@@ -17,8 +17,8 @@ from __future__ import print_function
 import numpy as np
 import itertools
 
-from ...mytypes import complexnp
-from ...mytypes import doublenp
+from ...wrappers.mytypes import complexnp
+from ...wrappers.mytypes import doublenp
 
 # Cython imports
