+++ qmeq/approach/elph/c_redfield.pxd
@@ -3,11 +3,11 @@ cimport numpy as np
 
 from ..c_aprclass cimport ApproachElPh
 
-from ..c_aprclass cimport bool_t
-from ..c_aprclass cimport int_t
-from ..c_aprclass cimport long_t
-from ..c_aprclass cimport double_t
-from ..c_aprclass cimport complex_t
+from ...wrappers.c_mytypes cimport bool_t
+from ...wrappers.c_mytypes cimport int_t
+from ...wrappers.c_mytypes cimport long_t
+from ...wrappers.c_mytypes cimport double_t
+from ...wrappers.c_mytypes cimport complex_t
 
 
 cdef class ApproachRedfield(ApproachElPh):
