+++ qmeq/specfunc/c_specfunc_elph.pxd
@@ -1,17 +1,20 @@
 import numpy as np
 cimport numpy as np
 
-ctypedef np.uint8_t bool_t
-ctypedef np.int_t int_t
-ctypedef np.int64_t long_t
-ctypedef np.float64_t double_t
-ctypedef np.complex128_t complex_t
+from ..wrappers.c_mytypes cimport bool_t
+from ..wrappers.c_mytypes cimport int_t
+from ..wrappers.c_mytypes cimport long_t
+from ..wrappers.c_mytypes cimport double_t
+from ..wrappers.c_mytypes cimport complex_t
+
 
 cdef class Func:
     cpdef double_t eval(self, double_t x)
 
+
 cdef class FuncBose:
     cdef double_t eval(self, double_t x)
+
 
 cdef class FuncPauliElPh:
     cdef Func dos
@@ -23,6 +26,7 @@ cdef class FuncPauliElPh:
     cdef bool_t bath_func_q
     cdef public double_t val
     cpdef void eval(self, double_t Ebbp, int_t l)
+
 
 cdef class Func1vNElPh:
     cdef Func dos
