+++ qmeq/wrappers/c_mytypes.pxd
@@ -1,0 +1,9 @@
+import numpy as np
+cimport numpy as np
+
+
+ctypedef np.uint8_t bool_t
+ctypedef np.int_t int_t
+ctypedef np.int64_t long_t
+ctypedef np.float64_t double_t
+ctypedef np.complex128_t complex_t
